+++ rust/alera-cli/src/cli.rs
@@ -25,11 +25,7 @@ pub enum Command {
     /// Run the persistent runtime host sidecar.
     #[command(name = RUNTIME_HOST_COMMAND)]
     RuntimeHost(TerminalHostArgs),
-
-    /// Serve newline-delimited runtime proxy requests over stdin/stdout.
-    #[command(name = "runtime-proxy")]
     RuntimeProxy,
-
     /// Run the persistent terminal host sidecar.
     #[command(name = TERMINAL_HOST_COMMAND)]
     TerminalHost(TerminalHostArgs),
